Fix Geocoder::Calculations.endpoint comment
I just spent a few hours trying to get this function to work, then realized I had my parameters in the wrong order. I was following the order listed in the comment, rather than the actual method definition! *facepalm* Anyhow, this fixes the comment, for the next person to read it.